+++ weave/legacy/artifact_fs.py
@@ -6,9 +6,9 @@ import json
 import os
 import typing
 
-from weave import errors, ref_base, ref_util
+from weave import errors, ref_util
 from weave import weave_types as types
-from weave.legacy import artifact_base, file_base, object_context, uris
+from weave.legacy import artifact_base, file_base, object_context, ref_base, uris
 from weave.legacy.language_features.tagging import tag_store
 
 if typing.TYPE_CHECKING:
