+++ weave/legacy/graph_mapper.py
@@ -1,4 +1,4 @@
-from weave import ref_base
+from weave.legacy import ref_base
 from weave import weave_types as types
 from weave.legacy import graph, mappers, node_ref
 from weave.legacy import mappers_python_def as mappers_python
