+++ weave/legacy/mappers_arrow.py
@@ -5,7 +5,7 @@ from contextlib import contextmanager
 
 import pyarrow as pa
 
-from weave import errors, ref_base
+from weave import errors
 from weave import weave_types as types
 from weave.legacy import (
     arrow_util,
@@ -15,6 +15,7 @@ from weave.legacy import (
     mappers_weave,
     node_ref,
     partial_object,
+    ref_base,
 )
 from weave.legacy import mappers_python_def as mappers_python
 from weave.legacy.arrow import arrow
