+++ weave/legacy/mappers_publisher.py
@@ -2,7 +2,7 @@ import copy
 import dataclasses
 import typing
 
-from weave import errors, ref_base, storage, weave_internal
+from weave import errors, storage, weave_internal
 from weave import weave_types as types
 from weave.legacy.node_ref import ref_to_node
 
@@ -15,6 +15,7 @@ from weave.legacy import (
     graph,
     mappers,
     mappers_python_def,
+    ref_base,
 )
 from weave.legacy.artifact_wandb import likely_commit_hash
 from weave.legacy.language_features.tagging import tag_store, tagged_value_type
