+++ weave/legacy/ops_primitives/artifacts.py
@@ -4,8 +4,8 @@ import pathlib
 import typing
 
 from weave.query_api import op
-from weave import ref_base, types
-from weave.legacy import artifact_fs
+from weave import types
+from weave.legacy import artifact_fs, ref_base
 from weave.legacy.artifact_local import WORKING_DIR_PREFIX, LocalArtifact
