+++ weave/legacy/ref_base.py
@@ -6,19 +6,18 @@ import weakref
 from typing import Sequence
 
 from weave.client_context import weave_client as weave_client_context
-from weave.legacy import box, context_state, object_context
+from weave.legacy import box, context_state, object_context, uris
 from weave.legacy.language_features.tagging import tag_store
 
-from . import errors
-from . import weave_types as types
-from .legacy import uris
+from .. import errors
+from .. import weave_types as types
 
 # We store Refs here if we can't attach them directly to the object
 REFS: weakref.WeakValueDictionary[int, "Ref"] = weakref.WeakValueDictionary()
 
 if typing.TYPE_CHECKING:
-    from . import weave_types as types
-    from .trace import weave_client
+    from .. import weave_types as types
+    from ..trace import weave_client
 
 
 def _map_to_ref_strs(obj: typing.Any) -> typing.Any:
