+++ weave/query_api.py
@@ -9,7 +9,7 @@ from weave.legacy.graph import Node
 from weave.legacy import graph_mapper as _graph_mapper
 
 from . import storage as _storage
-from . import ref_base as _ref_base
+from .legacy import ref_base as _ref_base
 from weave.legacy import wandb_api as _wandb_api
 
 from . import weave_internal as _weave_internal
