+++ weave/trace/weave_client.py
@@ -51,7 +51,7 @@ from weave.trace_server.trace_server_interface import (
 )
 
 if typing.TYPE_CHECKING:
-    from .. import ref_base
+    from ..legacy import ref_base
 
 
 # Controls if objects can have refs to projects not the WeaveClient project.
